Implemented OfferTable Function
Implemented a function imported and called in entry point of delivery cost calculator. Purpose: Makes changing, adding and removing test data super simple, just need to put it in 'criteria.txt' in the right format without having to interact with the code. The 'generateOfferTable' function takes care of the rest.